requirements.txt
Brotli==1.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
decorator==4.4.2
Flask==2.1.2
Flask-WTF==1.2.1
gunicorn==20.1.0
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mutagen==1.47.0
numpy==1.25.0
Pillow==9.5.0
proglog==0.1.10
pycryptodomex==3.19.0
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.65.0
urllib3==2.0.3
validators==0.20.0
websockets==11.0.3
Werkzeug==2.1.2
WTForms==3.0.1
youtube-dl==2021.12.17
yt-dlp==2023.9.24
zipp==3.15.0
